WIP for card pop view
