+++ scripts/card_launcher.js
@@ -4,21 +4,50 @@ var _ = require('underscore'),
   React = require('react'),
   MonkeyText = require('./monkey_text'),
   MonkeyImageText = require('./monkey_image_text'),
+  MonkeyImageTextEdit = require('./monkey_image_text_edit'),
   RB = require('react-bootstrap');
 
 var CardLauncher = React.createClass({
   render: function(){
-    if(this.props.card.type === 'MonkeyImageText'){
-      return (
-        <MonkeyImageText
-          card = {this.props.card} />
-      );
-    }else{
-      return (
-        <MonkeyText
-          card = {this.props.card} />
-      );
+    var component;
+
+    switch (this.props.type) {
+      case 'MonkeyImageText':
+        component = (
+          <MonkeyImageText
+            card = {this.props.card} />
+        );
+        break;
+      case 'MonkeyImageTextEdit':
+        component = (
+          <MonkeyImageTextEdit
+            isEditable = {true}
+            card = {this.props.card} />
+        );
+        break;
+      case 'MonkeyText':
+        component = (
+          <MonkeyText
+            card = {this.props.card} />
+        );
+        break;
+      case 'MonkeyTextEdit':
+        //todo, fix this
+        component = (
+          <MonkeyImageTextEdit
+            isEditable = {true}
+            card = {this.props.card} />
+        );
+        break;
+      default:
+        component = (
+          <MonkeyText
+            card = {this.props.card} />
+        );
+        break;
     }
+
+    return component;
   }
 });
